scripts/reddit_image_getter/requirements.txt
aiohttp==3.6.2
altgraph==0.17
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cfscrape==2.1.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
discord.py==1.4.1
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
flake8==3.8.3
Flask==1.0.2
Flask-Cors==3.0.6
flask-restplus==0.12.1
future==0.18.2
httmock==1.3.0
idna==2.9
isort==5.4.2
itsdangerous==0.24
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
pefile==2019.4.18
Pillow==7.2.0
puffotter==0.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.9.0
PyInstaller==3.6
pylint==2.6.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7.1
pyrsistent==0.18.0
python-dotenv==0.14.0
pytz==2020.1
pywin32-ctypes==0.2.0
rapidfuzz==0.7.10
requests==2.23.0
sentry-sdk==0.17.1
six==1.14.0
soupsieve==2.0
sty==1.0.0b12
tempora==3.0.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.35
websockets==8.1
Werkzeug==0.14.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.4.2
